cleaner test handling of phase flip; rename dss file with underscore
