+++ mne_sandbox/preprocessing/__init__.py
@@ -1,2 +1,2 @@
 from .eog import eog_regression
-from .dss import dss
+from ._dss import dss
